Fix: Replace Unsafe Java Deserialization with Jackson Mapper
This PR addresses a critical security vulnerability in the code that could lead to Remote Code Execution (RCE).

The code was using Java's built-in ObjectInputStream.readObject() for deserialization, which is known to be vulnerable to deserialization attacks if the input comes from an untrusted source.

This vulnerability was initially found and fixed in the git commit below.

References:
https://github.com/apache/helix/commit/7af17a31819859e6c46bcb3994ef7d52347760c4